SG Nuking Two-Step
Instead of having both a cut and nuke button show up on loading the
clip, now only a cut button shows up. Upon cutting the clip, a message
appears in the new status section. If you have the power to nuke clips,
a nuke button shows up with it. You can press that button, and the clip
will be nuked.